robotmk/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "robotmk"
authors = [{name = "Simon Meggle", email = "simon.meggle@elabit.de"}]
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
dynamic = ["version", "description"]
keywords = ["robotmk", "checkmk", "robotframework", "automation", "monitoring"]
readme = "README.md"
dependencies = [
    "click",
    "psutil",
    "apscheduler<=4",
    "pyyaml",
    "mergedeep",
    "robotframework",
    "loguru",
    "schema",
    "tabulate"
]

[project.scripts]
robotmk = "robotmk.cli.cli:main"
# TODO: add more scripts here
# robotmk-agent
# robotmk-specialagent (calls internally agent & output!)
# robotmk-output
